# docs/OpenAPI.yaml
openapi: 3.0.0
info:
  title: Todo app OpenAPI spec
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: ToDoApp server
paths:
...
  /v1/private/users:
    get:
      summary: Получение информации о пользователе
      tags:
        - Users
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Успешный ответ с информацией о пользователе
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "default":
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Обновление информации пользователя
      tags:
        - Users
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                  example: John
                  nullable: true
                  description: Может быть null, если не изменяется
                user_pass:
                  type: string
                  example: aaa123
                  nullable: true
                  description: Может быть null, если не изменяется
                user_email:
                  type: string
                  example: John@mail.go
                  nullable: true
                  description: Может быть null, если не изменяется
      responses:
        "200":
          description: Обновление информации пользователя прошло успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "default":
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Удаление пользователя
      tags:
        - Users
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Удаление прошло успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "default":
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Cookie-based аутентификация
  schemas:
...
    UserResponse:
      type: object
      properties:
        user_id:
          type: string
          example: b6609ddd-95f4-42f0-993e-7f07f3fa1d5b
        user_name:
          type: string
          example: John
        user_email:
          type: string
          example: John@mail.go
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: successful